packages/analyzer/requirements.txt
fastapi
uvicorn[standard]
psd-tools
pillow
python-multipart
pydantic
ezdxf
anthropic
matplotlib
python-dotenv
opencv-python-headless
numpy
